pyproject.toml
[project]
name = "hyperweave"
version = "0.1.0"
description = "Compositor API for self-contained SVG artifacts from semantic parameters."
requires-python = ">=3.12"
license = "MIT"
authors = [{ name = "InnerAura Labs" }]
readme = "README.md"
keywords = ["svg", "compositor", "badges", "artifacts", "mcp"]

dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "typer[all]>=0.15",
    "jinja2>=3.1",
    "httpx>=0.28",
    "fastmcp>=3.0",
    "pyyaml>=6.0",
]

[project.scripts]
hyperweave = "hyperweave.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hyperweave"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "pytest-mock>=3.14",
    "pytest-pretty>=1.2",
    "pytest-randomly>=3.16",
    "pytest-sugar>=1.0",
    "pytest-xdist>=3.5",
    "ruff>=0.9",
    "mypy>=1.14",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = ["--strict-markers", "--strict-config", "-ra"]
markers = [
    "slow: marks tests as slow",
    "snapshot: visual regression",
    "live: network access (deselect with '-m \"not live\"')",
]

[tool.ruff]
line-length = 120
target-version = "py312"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "TCH", "RUF"]

[tool.ruff.format]
docstring-code-format = true

[tool.mypy]
strict = true
plugins = ["pydantic.mypy"]
